TesteQualidade

+++ junit5-samples-/junit5-jupiter-starter-gradle/src/main/java/com/example/project/Calculator.java
@@ -11,19 +11,21 @@
 package com.example.project;
 
 public class Calculator {
-
-	public int soma (int a, int b) {
+	
+	public double soma (double a, double b) {
 		return a + b;
 	}
 	public double div (double a, double b){
 		return a / b;
 	}
 	
-	public int sub (int a, int b) {
+	public double sub (double a, double b) {
 		return a - b;
 	}
 	
-	public int mult (int a, int b) {
+	public double mult (double a, double b) {
 		return a * b;
 	}
+	
+	
 }
